+++ bc/recruitment/models.py
@@ -7,6 +7,7 @@ from django.db import models
 from django.db.models import Count, F
 from django.db.models.signals import pre_delete, pre_save
 from django.dispatch import receiver
+from django.http import Http404
 from django.shortcuts import get_object_or_404, render
 from django.template.loader import render_to_string
 from django.urls import reverse
@@ -19,6 +20,8 @@ from wagtail.core import blocks
 from wagtail.core.fields import StreamField
 from wagtail.images.edit_handlers import ImageChooserPanel
 from wagtail.search import index
+
+from wagtailorderable.models import Orderable
 
 from bc.utils.email import NotifyEmailMessage
 
@@ -49,10 +52,11 @@ class JobSubcategory(models.Model):
         ordering = ["title"]
 
 
-class JobCategory(models.Model):
+class JobCategory(Orderable, models.Model):
     title = models.CharField(max_length=128)
     description = models.TextField(blank=True)
     subcategories = models.ManyToManyField(JobSubcategory, related_name="categories")
+    is_schools_and_early_years = models.BooleanField(default=False)
 
     slug = models.SlugField(
         allow_unicode=True,
@@ -67,20 +71,37 @@ class JobCategory(models.Model):
     get_subcategories_list.short_description = "Subcategories"
 
     @staticmethod
-    def get_categories_summary():
+    def get_school_and_early_years_categories():
+        return list(
+            JobCategory.objects.filter(is_schools_and_early_years=True).values_list(
+                "slug", flat=True
+            )
+        )
+
+    @staticmethod
+    def get_categories_summary(queryset=None):
         """Returns a QuerySet that returns dictionaries, when used as an iterable.
 
            The dictionary keys are: category (category id), count, title, description
            This is ordered by highest count first.
         """
+        if not queryset:
+            queryset = TalentLinkJob.objects.all()
+
         job_categories = (
-            TalentLinkJob.objects.annotate(category=F("subcategory__categories"))
+            queryset.annotate(category=F("subcategory__categories"))
             .exclude(category=None)
             .values("category")
-            .annotate(id=F("subcategory__categories__slug"))
+            .annotate(key=F("subcategory__categories__slug"))
             .annotate(count=Count("category"))
             .annotate(label=F("subcategory__categories__title"))
             .annotate(description=F("subcategory__categories__description"))
+            .annotate(
+                is_schools_and_early_years=F(
+                    "subcategory__categories__is_schools_and_early_years"
+                )
+            )
+            .annotate(sort_order=F("subcategory__categories__sort_order"))
             .order_by("-count")
         )
 
@@ -165,6 +186,13 @@ class TalentLinkJob(models.Model):
         max_length=255, help_text="Salary group for filtering"
     )
     searchable_location = models.CharField(max_length=255)
+    location_postcode = models.CharField(max_length=8, blank=True)
+    location_lat = models.DecimalField(
+        max_digits=9, decimal_places=6, null=True, blank=True
+    )
+    location_lon = models.DecimalField(
+        max_digits=9, decimal_places=6, null=True, blank=True
+    )
     is_published = models.BooleanField(default=True)
     posting_start_date = models.DateTimeField()
     posting_end_date = models.DateTimeField()
@@ -271,11 +299,25 @@ class RecruitmentHomePage(RoutablePageMixin, BasePage):
     @route(r"^job_detail/(\d+)/$")
     def job_detail(self, request, talentlink_id):
         page = get_object_or_404(TalentLinkJob, talentlink_id=talentlink_id)
-        return render(request, "patterns/pages/jobs/job_detail.html", {"page": page})
+        return render(
+            request,
+            "patterns/pages/jobs/job_detail.html",
+            {"page": page, "show_apply_button": page.show_apply_button},
+        )
 
     @route(r"^apply/$")
     def apply(self, request):
-        return render(request, "patterns/pages/jobs/apply.html")
+        job_id = request.GET.get("jobId")
+        if job_id:
+            talentlink_id = job_id.split("-")[1]
+            page = get_object_or_404(TalentLinkJob, talentlink_id=talentlink_id)
+        else:
+            raise Http404("Missing job details")
+        return render(
+            request,
+            "patterns/pages/jobs/apply.html",
+            {"page": page, "show_apply_button": False},
+        )
 
 
 class RecruitmentIndexPage(BasePage):