+++ bc/recruitment/utils.py
@@ -1,9 +1,13 @@
 import json
 
 from django.contrib.postgres.search import SearchQuery, SearchRank, SearchVector
+from django.db.models import F
+from django.db.models.functions import ACos, Cos, Radians, Sin
+
+import requests
 
 from bc.recruitment.constants import JOB_FILTERS
-from bc.recruitment.models import RecruitmentHomePage, TalentLinkJob
+from bc.recruitment.models import JobCategory, RecruitmentHomePage, TalentLinkJob
 
 
 def is_recruitment_site(request):
@@ -17,7 +21,10 @@ def get_current_search(querydict):
     search = {}
 
     if querydict.get("query", None):
-        search["query"] = querydict.get("query", None)
+        search["query"] = querydict["query"]
+
+    if querydict.get("postcode", None):
+        search["postcode"] = querydict["postcode"]
 
     # Loop through our filters so we don't just store any query params
     for filter in JOB_FILTERS:
@@ -32,11 +39,13 @@ def get_current_search(querydict):
 def get_job_search_results(querydict, queryset=None):
     if queryset is None:
         queryset = TalentLinkJob.objects.all()
+
     search_query = querydict.get("query", None)
 
     if search_query:
         vector = (
             SearchVector("title", weight="A")
+            + SearchVector("job_number", weight="A")
             # + SearchVector("short_description", weight="A")
             + SearchVector("searchable_location", weight="B")
             + SearchVector("description", weight="C")
@@ -52,6 +61,15 @@ def get_job_search_results(querydict, queryset=None):
         # Order by newest job at top
         search_results = queryset.order_by("posting_start_date")
 
+    # Process 'hide schools and early years job'
+    if querydict.get("hide_schools_and_early_years", False):
+        schools_and_early_years_categories = (
+            JobCategory.get_school_and_early_years_categories()
+        )
+        search_results = search_results.exclude(
+            subcategory__categories__slug__in=schools_and_early_years_categories
+        )
+
     # Process filters
     for filter in JOB_FILTERS:
         # QueryDict.update() used in send_job_alerts.py adds the values as list instead of multivalue dict.
@@ -64,7 +82,60 @@ def get_job_search_results(querydict, queryset=None):
 
         if selected:
             search_results = search_results.filter(
-                **{filter["filter_key"] + "__in": selected}
+                **{
+                    filter["filter_key"] + "__in": selected
+                }  # TODO: make case insensitive
             )
 
+    # Process postcode search
+    search_postcode = querydict.get("postcode", None)
+    if search_postcode:
+        postcode_response = requests.get(
+            "https://api.postcodes.io/postcodes/" + search_postcode
+        )
+        if postcode_response.status_code == 200:
+            postcode_response_json = postcode_response.json()
+            search_lon = postcode_response_json["result"]["longitude"]
+            search_lat = postcode_response_json["result"]["latitude"]
+
+            search_results = search_results.annotate(
+                distance=GetDistance(search_lat, search_lon)
+            ).order_by("distance")
+
+            if search_query:
+                # Rank is only used when there is a search query
+                search_results = search_results.order_by("distance", "-rank")
+
     return search_results
+
+
+def GetDistance(point_latitude, point_longitude):
+    # Calculate distance. See https://www.thutat.com/web/en/programming-and-tech-stuff/
+    # web-programming/postgres-query-with-gps-distance-calculations-without-postgis/
+    distance = (
+        ACos(
+            Sin(Radians(F("location_lat"))) * Sin(Radians(point_latitude))
+            + Cos(Radians(F("location_lat")))
+            * Cos(Radians(point_latitude))
+            * Cos(Radians(F("location_lon") - point_longitude))
+        )
+        * 6371
+        * 1000
+    )
+
+    return distance
+
+
+def get_school_and_early_years_count(search_results=None):
+    if search_results is None:
+        search_results = TalentLinkJob.objects.all()
+
+    schools_and_early_years_categories = (
+        JobCategory.get_school_and_early_years_categories()
+    )
+    if len(schools_and_early_years_categories):
+        search_results = search_results.filter(
+            subcategory__categories__slug__in=schools_and_early_years_categories
+        )
+
+    return len(search_results)