+++ bc/recruitment/wagtail_hooks.py
@@ -20,6 +20,7 @@ class TalentLinkJobModelAdmin(ModelAdmin):
         "title",
         "get_categories_list",
         "last_imported",
+        "location_postcode",
     )
 
     def job_link(self, obj):